+++ src/routing/route/definition/create-route-definition.ts
@@ -4,7 +4,7 @@ import type { RouteOptions } from '@/routing/route/route-options';
 import type { RouterMethod } from '@/routing/route/router-method.type';
 import type { HttpMethod } from '@/routing/verb/http-method.type';
 import type { Request } from 'koa';
-import type { RouteDefinition } from './route-definition';
+import type { RouteDefinition } from './route-definition.type';
 
 interface RouteDefinitionInput<TRequest extends Request = HttpRequest> {
   name?: string;
