+++ src/routing/route/definition/resolve-route-options.ts
@@ -1,6 +1,6 @@
 import type { RouteOptions } from '@/routing/route/route-options';
 import type { Request } from 'koa';
-import type { RouteDefinition } from './route-definition';
+import type { RouteDefinition } from './route-definition.type';
 
 export function resolveRouteOptions(options: RouteOptions): Pick<RouteDefinition, 'parseBody' | 'bodyOptions'> {
   return {
