+++ src/routing/route/source/route-source.ts
@@ -1,6 +1,6 @@
 import type { HttpRequest } from '@/Http';
 import type { RouteGroupDefinition } from '@/routing/group/route-group';
-import type { RouteDefinition } from '@/routing/route/definition/route-definition';
+import type { RouteDefinition } from '@/routing/route/definition/route-definition.type';
 import type { Request } from 'koa';
 
 export type RouteSource<TRequest extends Request = HttpRequest> =
